+++ hub-client/src/model/yivi/Tyivi.ts
@@ -1,0 +1,6 @@
+type Attribute = {
+	attribute: string;
+	label?: string;
+};
+
+export { Attribute };