+++ CHANGELOG.md
@@ -11,6 +11,9 @@ _Please add a brief description of any changes and any migrations to be performe
 
 ## Changed not yet merged to stable
 
+- [NEW] Added the option to remove attribute values from existing secured rooms.
+- [NEW] Updated validation for the form that is used to edit rooms.
+- [NEW] Added user notifications for important events related to the secured rooms.
 - [BREAKING] Added `phc.global_client_url` to `pubhubs.toml`.
 
 ## 10 July 2025 - v2.2.0